Split the Pokemon`s Test to API and Web
